Move version into Issue
