+++ fronts-client/src/bundles/editionsIssueBundle.ts
@@ -2,6 +2,8 @@ import createAsyncResourceBundle, {
   State,
 } from 'lib/createAsyncResourceBundle';
 import { EditionsIssue } from 'types/Edition';
+import { ThunkResult, Dispatch } from 'types/Store';
+import { getIssueSummary } from 'services/editionsApi';
 
 export const {
   actions,
@@ -17,3 +19,15 @@ export const {
 export type EditionsIssueState = State<EditionsIssue>;
 
 export default reducer;
+
+export const getEditionIssue = (
+  id: string
+): ThunkResult<Promise<void>> => async (dispatch: Dispatch) => {
+  try {
+    dispatch(actions.fetchStart());
+    const issue = await getIssueSummary(id);
+    dispatch(actions.fetchSuccess(issue));
+  } catch (error) {
+    dispatch(actions.fetchError('Failed to get issue'));
+  }
+}
